move loading card to a common component
