+++ frontend/src/app/accountCreation/PasswordForm/PasswordForm.tsx
@@ -12,7 +12,7 @@ import {
 } from "../../utils/text";
 import { AccountCreationApi } from "../AccountCreationApiService";
 import { RootState } from "../../store";
-import { LoadingCard } from "../LoadingCard/LoadingCard";
+import { LoadingCard } from "../../commonComponents/LoadingCard/LoadingCard";
 import { Card } from "../../commonComponents/Card/Card";
 import { CardBackground } from "../../commonComponents/CardBackground/CardBackground";
 import StepIndicator from "../../commonComponents/StepIndicator";
