+++ frontend/src/app/commonComponents/LoadingCard/LoadingCard.stories.tsx
